Group meson commands in preparation for other build systems
