+++ devpy/cmds/__init__.py
@@ -1,3 +1,8 @@
-from ._build import build
-from ._test import test
-from ._shell import ipython, python, shell
+from . import meson
+
+# Backward compatibility with older versions
+build = meson.build
+test = meson.test
+ipython = meson.ipython
+python = meson.python
+shell = meson.shell
