+++ devpy/cmds/meson.build
@@ -1,8 +1,7 @@
 python_sources = [
   '__init__.py',
-  '_build.py',
-  '_shell.py',
-  '_test.py',
+  'build.py',
+  'meson.py',
   'util.py'
 ]
