+++ devpy/meson.build
@@ -1,6 +1,8 @@
 python_sources = [
   '__init__.py',
-  '__main__.py'
+  '__main__.py',
+  'color_format.py',
+  'sectioned_help.py'
 ]
 
 py.install_sources(
